movimentation

+++ game/src/com/br/enginejava/entidades/Entity.java
@@ -5,8 +5,8 @@ import java.awt.image.BufferedImage;
 
 public class Entity {
 	
-	protected int x;
-	protected int y;
+	protected double x;
+	protected double y;
 	protected int width;
 	protected int height;
 	protected BufferedImage sprite;
@@ -20,7 +20,7 @@ public class Entity {
 	}
 
 	public int getX() {
-		return x;
+		return (int)x;
 	}
 
 	public void setX(int x) {
@@ -28,7 +28,7 @@ public class Entity {
 	}
 
 	public int getY() {
-		return y;
+		return (int)y;
 	}
 
 	public void setY(int y) {
